Added new `useCallingContext` configuration boolean
The `useCallingContext` configuration (defaulted to false) causes any functions that provide a `computer` function at the call site to add the caller's `coroutineContext` to the scope used to perform the computation.

This change means contextual information such as ThreadContextElements to not be lost when those computations are run.

This applies to both `Cache` and `LoadingCache` `compute` functions that are provided at the call site only.

In addition to this change the following was added:

* `java-test-fixtures` plugin to include a `BooleanThreadContextElement` type for unit tests
* Unit tests added for all functions that provide `compute` at the call site to verify `useCallingContext = true` preserves the calling context

Fixes #11