fix loading cache nullability (#50)
Caffeine permits a loader to return null, however the recent alignment to JSpecify annotations applied the NonNull to the loader lambda return type making it impossible to provide a loader that returns null.

This attempts to fix the nullability by enforcing `& Any` only where it's required to match the JSpecify annotations.

One issue encountered was with bulk loaders, if the bulk mapping function returns a map that doesn't contain the requested key, the cache _will_ return `null` on a single `get(K)`. As such, I've ensured that `asBulkLoadingCache` returns a `LoadingCache<K, V?>`